+++ CHANGELOG.md
@@ -8,6 +8,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-952 Add meta "description" to the index.html
 - YLP-972 Add United Kingdom
 ### Changed
+- YLP-955 Sign-up consent next button should be greyed when boxes are not checked
 - YLP-958 Review text justification in modal windows
 
 ## 2.0.3 - 2021-08-20